Implement LZ4 wrapper, replacing LZW compression

+++ include/taproot/log.h
@@ -39,7 +39,7 @@ class Log {
         std::string getEarliestSnapshot();
         void rotateLogs();
         void compactLog(const std::unordered_map<std::string, std::string>& state, const bool hasBeenAltered);
-        std::vector<uint16_t> compress(const std::vector<uint8_t>& input);
+        std::vector<uint8_t> compress(const std::vector<uint8_t>& input);
 
         void closeLog();
 };
